requirement.txt
pandas==2.2.2
requests==2.31.0
google-cloud-bigquery==3.17.2
google-cloud-storage==2.14.0
